refac: Move PoC config to configurables
