+++ UMA-Bots/Large-Relayer-Token-Balance-Change/src/configurables.ts
@@ -43,3 +43,16 @@ export const MAINNET_MONITORED_ADDRESSES = [
 export const OPTIMISM_MONITORED_ADDRESSES = [];
 export const POLYGON_MONITORED_ADDRESSES = [];
 export const ARBITRUM_MONITORED_ADDRESSES = [];
+
+// List of PoC tokens and monitored addresses
+
+export const GOERLI_MONITORED_ADDRESSES = [
+  "0x1Abf3a6C41035C1d2A3c74ec22405B54450f5e13",
+  "0x628bfE54739098012bDc282EFA2F74c226FF5d40",
+];
+
+export const GOERLI_MONITORED_TOKENS = [
+  "0x6b79C63a7d3ACF14272eB19B384046bAd9a3C6E2", // USDT
+];
+
+export const GOERLI_PERCENTAGE_CHANGE_THRESHOLD = 0;
